EMR types restructure for Symptoms, Diagnosis, Allergy (#9705)
