+++ src/components/Questionnaire/structured/types.ts
@@ -6,13 +6,13 @@ import {
 import {
   AllergyIntolerance,
   AllergyIntoleranceRequest,
-} from "@/types/emr/allergyIntolerance";
+} from "@/types/emr/allergyIntolerance/allergyIntolerance";
+import { Diagnosis, DiagnosisRequest } from "@/types/emr/diagnosis/diagnosis";
 import { Encounter, EncounterEditRequest } from "@/types/emr/encounter";
 import { MedicationRequest } from "@/types/emr/medicationRequest";
 import { MedicationStatement } from "@/types/emr/medicationStatement";
-import { Diagnosis, DiagnosisRequest } from "@/types/questionnaire/diagnosis";
+import { Symptom, SymptomRequest } from "@/types/emr/symptom/symptom";
 import { StructuredQuestionType } from "@/types/questionnaire/question";
-import { Symptom, SymptomRequest } from "@/types/questionnaire/symptom";
 
 // Map structured types to their data types
 export interface StructuredDataMap {
