+++ src/types/emr/symptom/symptom.ts
@@ -1,5 +1,5 @@
-import { UserBase } from "../user/user";
-import { Code } from "./code";
+import { Code } from "../../questionnaire/code";
+import { UserBase } from "../../user/user";
 
 export const SYMPTOM_CLINICAL_STATUS = [
   "active",
