+++ src/types/questionnaire/form.ts
@@ -1,15 +1,14 @@
 import { FollowUpAppointmentRequest } from "@/components/Schedule/types";
 
-import { AllergyIntolerance } from "@/types/emr/allergyIntolerance";
+import { AllergyIntolerance } from "@/types/emr/allergyIntolerance/allergyIntolerance";
+import { Diagnosis } from "@/types/emr/diagnosis/diagnosis";
 import { Encounter } from "@/types/emr/encounter";
 import { MedicationRequest } from "@/types/emr/medicationRequest";
 import { MedicationStatement } from "@/types/emr/medicationStatement";
+import { Symptom } from "@/types/emr/symptom/symptom";
 import { Code } from "@/types/questionnaire/code";
 import { Quantity } from "@/types/questionnaire/quantity";
 import { StructuredQuestionType } from "@/types/questionnaire/question";
-
-import { Diagnosis } from "./diagnosis";
-import { Symptom } from "./symptom";
 
 export type ResponseValue = {
   type:
